More old-style cast cleanups.

+++ src/doom/p_inter.cpp
@@ -772,7 +772,7 @@ P_KillMobj
     // [crispy] randomly flip corpse, blood and death animation sprites
     if (target->flags & MF_FLIPPABLE)
     {
-	target->health = (target->health & (int)~1) - (Crispy_Random() & 1);
+	target->health = (target->health & ~1) - (Crispy_Random() & 1);
     }
 
     if (target->tics < 1)
